♻️(frontend) move ContentSection to prevent confusion with Content pages
Since ContentSection is used only on the landing page blocks I felt
moving it out of the layout folder, now that we have a "ContentPage"
component that is rather unrelated, made things clearer.
