+++ src/components/Layout/index.ts
@@ -2,5 +2,4 @@ export * from "./Footer";
 export * from "./NavBar";
 export * from "./MetaHeader";
 export * from "./Layout";
-export * from "./ContentSection";
 export * from "./ContentPage";
