+++ src/components/sections/index.tsx
@@ -1,0 +1,1 @@
+export * from "./ContentSection";
